backend/auth-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.2.2
sqlalchemy[asyncio]==2.0.49
asyncpg==0.31.0
PyJWT==2.12.1
google-auth==2.30.0
